pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.11"
django-environ = "^0.12.1"
dj-database-url = "^3.1.1"
djangorestframework-simplejwt = "^5.5.1"
celery = "^5.6.2"
django-filter = "^25.2"
coverage = "^7.13.4"
drf-yasg = "^1.21.14"
requests = "^2.32.5"
django-cors-headers = "^4.9.0"
pillow = "^12.1.1"
eventlet = "^0.40.4"
redis = "^7.2.0"
django-celery-results = "^2.6.0"
python-dotenv = "^1.2.1"
django = ">=4.2,<6.0"
django-celery-beat = "^2.8.1"
flake8 = "^7.3.0"
black = "^26.1.0"
mypy = "^1.19.1"
isort = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
